refactor(list): split init.lua into keymaps and autocmds sub-modules
